LayoutTests/fast/events/ios/rotation/basic-rotation-expected.txt

Before rotation
PASS window.innerWidth is 320
PASS window.innerHeight is 548
In orientationchange event handler:
FAIL window.innerWidth should be 568. Was 320.
FAIL window.innerHeight should be 320. Was 548.

In resize event handler:
FAIL window.innerWidth should be 568. Was 320.
PASS window.innerHeight is 320

After rotation
PASS window.innerWidth is 568
PASS window.innerHeight is 320
Rotation test.
